backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv

langgraph
langchain
langchain-openai
